\name{Rslippy-package}
\alias{Rslippy-package}
\alias{Rslippy}
\docType{package}
\title{
Generate Leaflet slippy maps from spatial objects.
}
\description{
This package renders tiles from Raster or Spatial* objects, and produces a HTML page that uses the Leaflet Javascript library to render them as a slippy map in a browser.
}
\details{
\tabular{ll}{
Package: \tab Rslippy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-21\cr
License: \tab What license is it under?\cr
}
The main functions are slippyRaster, for raster objects, and slippySpatial for Spatial* vector objects.
}
\author{
Grant Williamson
Maintainer: Grant Williamson <grant.williamson@utas.edu.au>
}
\keyword{ package }
\keyword{ Rslippy }
\keyword{ slippy }
\keyword{ raster }
\keyword{ spatial }
\keyword{ map }
\keyword{ leaflet }
